@@ -15,7 +15,7 @@
   ],
   "minimum-stability": "dev",
   "require": {
-    "php": "^7.3",
+    "php": "^7.4",
     "zendframework/zend-mvc": "^3.0.1",
     "zendframework/zend-mvc-form": "^1.0.0",
     "solarium/solarium": "5.*",
